+++ common/src/main/java/dev/xylonity/companions/config/SpawnConfig.java
@@ -3,6 +3,7 @@ package dev.xylonity.companions.config;
 import net.minecraft.core.Holder;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
+import dev.xylonity.knightlib.api.util.ResourceLocations;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.biome.Biome;
@@ -38,10 +39,10 @@ public class SpawnConfig {
         for (int i = 3; i < parts.length; i++) {
             String part = parts[i];
             if (part.startsWith("#")) {
-                ResourceLocation tagId = new ResourceLocation(part.substring(1));
+                ResourceLocation tagId = ResourceLocations.parse(part.substring(1));
                 tagList.add(TagKey.create(Registries.BIOME, tagId));
             } else {
-                biomeList.add(new ResourceLocation(part));
+                biomeList.add(ResourceLocations.parse(part));
             }
         }